backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
backports.tarfile==1.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-log==0.4.0
colorama==0.4.6
cryptography==46.0.2
Deprecated==1.2.18
deprecation==2.1.0
docutils==0.22.2
dotty-dict==1.3.1
Flask==2.3.3
Flask-Cors==3.0.10
Flask-Limiter==4.0.0
Flask-Mail==0.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.29.0
google-auth-oauthlib==1.2.2
gotrue==1.3.1
gspread==5.12.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
importlib_metadata==8.7.0
invoke==1.7.3
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
Jinja2==3.1.6
keyring==25.6.0
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
nh3==0.3.1
oauthlib==3.3.1
ordered-set==4.1.0
packaging==25.0
pkginfo==1.12.1.2
postgrest==0.10.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gitlab==3.15.0
python-semantic-release==7.33.2
pywin32-ctypes==0.2.3
readme_renderer==44.0
realtime==1.0.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.2.0
rsa==4.9.1
SecretStorage==3.4.0
semver==2.13.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
tomlkit==0.13.3
tqdm==4.67.1
twine==3.8.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
